requirements.txt
requests
qrcode
pillow
discord.py